+++ CHANGELOG.md
@@ -5,6 +5,42 @@ Questions and feedback can be submitted to the Electric Power Research Institute
 
 The format is based on [Keep a Changelog] (https://keepachangelog.com/en/1.0.0/).
 
+## [1.2.0] - 2021-09-08 to 2022-03-30
+### Added
+- pytests to ensure that the default model parameter CSV file runs when input into run_StorageVET.py
+- pytests to ensure that with each technology active along with a battery, the default model parameter CSV runs
+- adds three new Finance tag inputs to set the fuel price: liquid, gas, other
+- allow the forces use of the GLPK_MI solver when necessary
+- a copy of the model parameters input CSV is now copied to the Results folder for each run
+- adds required rows for all technologies in the default model parameter csv
+
+### Changed
+- upgrade supported/recommended Python version to 3.8.13
+  - Python package requirements have been updated
+  - Update installation instructions: Python environment creation, conda-route, pip-route
+- Adds missing udis term in proforma of technologies for variable om cost and fuel cost
+- limit MACRS term to no greater than 20
+- re-structures how fuel costs are handled
+  - fuel costs are now constant over time (time independent)
+  - fuel prices are now set in a project-wide manner
+  - fuel cost is now determined by the fuel_type within a technology and an associated fuel_price in the Finance tag
+  - units of fuel are now relative to Btu and not gallons
+  - fuel price is collected as a user input in $/MMBtu but converted into $/kW upon ingest
+- changes all kw (lower-case w) labels to kW (upper-case W)
+- changes all BTU labels to Btu
+
+### Removed
+- remove fuel costs that are tied to a specific technology
+  - dropped use of natural gas and diesel labels for fuel
+
+### Fixed
+- corrects CAES to bypass degradation module
+- add fuel costs to proforma for CAES
+- fix bug where CAES was having its tag overwritten
+- MarketServiceUp calculates worst case energy based on dt instead of duration
+- FR energy throughput now divides by the ES rte in objective function and proforma
+- Enable user service to apply min discharge and min charge constraints
+
 ## [1.1.3] - 2021-08-04 to 2021-09-07
 ### Changed
 - Changed the expected type to float for yearly_degrade battery input